Move EnumConverters.cs from "Painting" to "DataTypes".
